UploadFile엔티티에 FK User 추가

+++ SpringBoot/src/main/java/com/seungwook/ktsp/domain/file/service/FileService.java
@@ -3,11 +3,16 @@ package com.seungwook.ktsp.domain.file.service;
 import com.seungwook.ktsp.domain.file.dto.AttachedFile;
 import com.seungwook.ktsp.domain.file.dto.AttachedFileInfo;
 import com.seungwook.ktsp.domain.file.entity.UploadFile;
+import com.seungwook.ktsp.domain.file.exception.FileException;
 import com.seungwook.ktsp.domain.file.service.domain.BoardFileDomainService;
 import com.seungwook.ktsp.domain.file.service.domain.UploadFileDomainService;
 import com.seungwook.ktsp.domain.file.service.policy.FileStoreService;
+import com.seungwook.ktsp.domain.user.entity.User;
+import com.seungwook.ktsp.domain.user.entity.enums.UserRole;
+import com.seungwook.ktsp.domain.user.service.UserDomainService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -26,13 +31,17 @@ public class FileService {
     private final FileStoreService fileStoreService;
     private final BoardFileDomainService boardFileDomainService;
     private final UploadFileDomainService uploadFileDomainService;
+    private final UserDomainService userDomainService;
 
     // 파일 업로드
     @Transactional
-    public String uploadFile(MultipartFile file, boolean isImageFile) {
+    public String uploadFile(long userId, MultipartFile file, boolean isImageFile) {
+
+        // 프록시 객체
+        User user = userDomainService.getReferenceById(userId);
 
         // 파일 저장
-        UploadFile uploadFile = fileStoreService.storeFile(file, isImageFile);
+        UploadFile uploadFile = fileStoreService.storeFile(user, file, isImageFile);
         uploadFileDomainService.save(uploadFile);
 
         // 파일 접근 경로 리턴
@@ -75,14 +84,20 @@ public class FileService {
 
     // 파일 삭제
     @Transactional
-    public void deleteFile(String uuid) {
+    public void deleteFile(String uuid, long requestUserId, UserRole requestUserRole) {
 
         UploadFile uploadFile = uploadFileDomainService.findByUuid(uuid);
+
+        // 파일 삭제 권한이 있는지 검사
+        boolean isNotResourceOwner = uploadFile.getUser().getId() != requestUserId;
+        boolean isNotAdmin = !requestUserRole.equals(UserRole.ADMIN);
+        if (isNotResourceOwner && isNotAdmin)
+            throw new FileException(HttpStatus.FORBIDDEN, "파일을 삭제할 권한이 없습니다.");
 
         // DB에서 제거
         uploadFileDomainService.delete(uploadFile);
 
-        // 스토리지에서 파일 삭제
+        // 스토리지에서 삭제
         fileStoreService.deleteFile(uploadFile);
     }
 }
